refactor(api): reorganize imports and remove unused action-policy file
