+++ apps/web/src/app/[lang]/(main)/(core)/management/identity/claim-types/_components/claim-types-table-data.tsx
@@ -9,8 +9,8 @@ import {tanstackTableCreateColumnsByRowData} from "@repo/ayasofyazilim-ui/molecu
 import type {Policy} from "@repo/utils/policies";
 import {Plus} from "lucide-react";
 import type {AppRouterInstance} from "next/dist/shared/lib/app-router-context.shared-runtime";
+import {isActionGranted} from "@repo/utils/policies";
 import type {IdentityServiceResource} from "src/language-data/core/IdentityService";
-import isActionGranted from "src/utils/page-policy/action-policy";
 
 type ClaimTypesTable = TanstackTableCreationProps<Volo_Abp_Identity_ClaimTypeDto>;
