+++ apps/web/src/app/[lang]/(main)/(core)/management/identity/users/[userId]/_components/form.tsx
@@ -12,7 +12,7 @@ import ConfirmDialog from "@repo/ayasofyazilim-ui/molecules/confirm-dialog";
 import {SchemaForm} from "@repo/ayasofyazilim-ui/organisms/schema-form";
 import {createUiSchemaWithResource} from "@repo/ayasofyazilim-ui/organisms/schema-form/utils";
 import {CustomMultiSelectWidget} from "@repo/ayasofyazilim-ui/organisms/schema-form/widgets";
-import {useGrantedPolicies} from "@repo/utils/policies";
+import {useGrantedPolicies, isActionGranted} from "@repo/utils/policies";
 import {Trash2} from "lucide-react";
 import {useRouter} from "next/navigation";
 import {useTransition} from "react";
@@ -20,7 +20,6 @@ import {handleDeleteResponse, handlePutResponse} from "@repo/utils/api";
 import {deleteUserByIdApi} from "src/actions/core/IdentityService/delete-actions";
 import {putUserApi} from "src/actions/core/IdentityService/put-actions";
 import type {IdentityServiceResource} from "src/language-data/core/IdentityService";
-import isActionGranted from "src/utils/page-policy/action-policy";
 
 type UserFormDto = Volo_Abp_Identity_IdentityUserDto & {
   organizationUnitIds?: string[] | null;
