+++ apps/web/src/app/[lang]/(main)/(core)/management/language-management/languages/[languageId]/_components/form.tsx
@@ -9,7 +9,7 @@ import {ActionList} from "@repo/ayasofyazilim-ui/molecules/action-button";
 import ConfirmDialog from "@repo/ayasofyazilim-ui/molecules/confirm-dialog";
 import {SchemaForm} from "@repo/ayasofyazilim-ui/organisms/schema-form";
 import {createUiSchemaWithResource} from "@repo/ayasofyazilim-ui/organisms/schema-form/utils";
-import {useGrantedPolicies} from "@repo/utils/policies";
+import {useGrantedPolicies, isActionGranted} from "@repo/utils/policies";
 import {Trash2} from "lucide-react";
 import {useRouter} from "next/navigation";
 import {useTransition} from "react";
@@ -17,7 +17,6 @@ import {handleDeleteResponse, handlePutResponse} from "@repo/utils/api";
 import {deleteLanguageByIdApi} from "src/actions/core/AdministrationService/delete-actions";
 import {putLanguageApi} from "src/actions/core/AdministrationService/put-actions";
 import type {AdministrationServiceResource} from "src/language-data/core/AdministrationService";
-import isActionGranted from "src/utils/page-policy/action-policy";
 
 export default function Form({
   languageData,
