+++ apps/web/src/app/[lang]/(main)/(core)/management/openiddict/applications/[applicationId]/_components/form.tsx
@@ -17,7 +17,7 @@ import AutoForm, {
   CustomCombobox,
   DependencyType,
 } from "@repo/ayasofyazilim-ui/organisms/auto-form";
-import {useGrantedPolicies} from "@repo/utils/policies";
+import {useGrantedPolicies, isActionGranted} from "@repo/utils/policies";
 import {Trash2} from "lucide-react";
 import {useParams, useRouter} from "next/navigation";
 import {useTransition} from "react";
@@ -25,7 +25,6 @@ import {handleDeleteResponse, handlePutResponse} from "@repo/utils/api";
 import {deleteApplicationByIdApi} from "src/actions/core/IdentityService/delete-actions";
 import {putApplicationApi} from "src/actions/core/IdentityService/put-actions";
 import type {IdentityServiceResource} from "src/language-data/core/IdentityService";
-import isActionGranted from "src/utils/page-policy/action-policy";
 
 interface FieldProps {
   field: {
