+++ apps/web/src/app/[lang]/(main)/(core)/management/openiddict/scopes/[scopeId]/_components/form.tsx
@@ -6,7 +6,7 @@ import {ActionList} from "@repo/ayasofyazilim-ui/molecules/action-button";
 import ConfirmDialog from "@repo/ayasofyazilim-ui/molecules/confirm-dialog";
 import {SchemaForm} from "@repo/ayasofyazilim-ui/organisms/schema-form";
 import {createUiSchemaWithResource} from "@repo/ayasofyazilim-ui/organisms/schema-form/utils";
-import {useGrantedPolicies} from "@repo/utils/policies";
+import {useGrantedPolicies, isActionGranted} from "@repo/utils/policies";
 import {Trash2} from "lucide-react";
 import {useRouter} from "next/navigation";
 import {useTransition} from "react";
@@ -14,7 +14,6 @@ import {handleDeleteResponse, handlePutResponse} from "@repo/utils/api";
 import {deleteScopeByIdApi} from "src/actions/core/IdentityService/delete-actions";
 import {putScopeApi} from "src/actions/core/IdentityService/put-actions";
 import type {IdentityServiceResource} from "src/language-data/core/IdentityService";
-import isActionGranted from "src/utils/page-policy/action-policy";
 
 export default function Form({
   languageData,
