+++ packages/utils/api/client-utils.ts
@@ -1,10 +1,10 @@
 "use client";
-import { toast } from "@/components/ui/sonner";
-import type { AppRouterInstance } from "next/dist/shared/lib/app-router-context.shared-runtime";
-import { ServerResponse } from "./types";
+import {toast} from "@/components/ui/sonner";
+import type {AppRouterInstance} from "next/dist/shared/lib/app-router-context.shared-runtime";
+import {ServerResponse} from "./types";
 
 export const handlePutResponse = (
-  response: { type: "success" | "api-error"; message: string },
+  response: {type: "success" | "api-error"; message: string},
   router?: AppRouterInstance,
   redirectTo?: string,
 ) => {
@@ -23,9 +23,7 @@ export const handlePutResponse = (
 export const handlePostResponse = <T>(
   response: ServerResponse<T>,
   router?: AppRouterInstance,
-  redirectTo?:
-    | string
-    | { prefix: string; identifier: keyof T; suffix?: string },
+  redirectTo?: string | {prefix: string; identifier: keyof T; suffix?: string},
 ) => {
   if (response.type === "success") {
     toast.success("Created successfully");
@@ -33,7 +31,7 @@ export const handlePostResponse = <T>(
     if (typeof redirectTo === "string") {
       router.push(redirectTo);
     } else if (redirectTo) {
-      const { prefix, suffix, identifier } = redirectTo;
+      const {prefix, suffix, identifier} = redirectTo;
       const id = (response.data[identifier] as string).toString();
       router.push(`${prefix}/${id}/${suffix}`);
     }
@@ -43,7 +41,7 @@ export const handlePostResponse = <T>(
   }
 };
 export const handleDeleteResponse = (
-  response: { type: "success" | "error" | "api-error"; message: string },
+  response: {type: "success" | "error" | "api-error"; message: string},
   router?: AppRouterInstance,
   redirectTo?: string,
 ) => {
@@ -58,10 +56,7 @@ export const handleDeleteResponse = (
     toast.error(response.message);
   }
 };
-export const handleGetResponseError = (response: {
-  type: "success" | "error" | "api-error";
-  message: string;
-}) => {
+export const handleGetResponseError = (response: {type: "success" | "error" | "api-error"; message: string}) => {
   if (response.type !== "success") {
     toast.error(response.message);
   }
