+++ packages/utils/auth/auth.ts
@@ -1,17 +1,13 @@
 import Credentials from "next-auth/providers/credentials";
 export type Awaitable<T> = T | PromiseLike<T>;
 
-import { AdapterUser } from "@auth/core/adapters";
-import {
-  fetchNewAccessTokenByRefreshToken,
-  fetchToken,
-  getUserData,
-} from "./auth-actions";
-import NextAuth, { AuthError } from "next-auth";
-import { MyUser } from "./auth-types";
-import { GetApiAbpApplicationConfigurationResponse } from "@ayasofyazilim/core-saas/AccountService";
+import {AdapterUser} from "@auth/core/adapters";
+import {fetchNewAccessTokenByRefreshToken, fetchToken, getUserData} from "./auth-actions";
+import NextAuth, {AuthError} from "next-auth";
+import {MyUser} from "./auth-types";
+import {GetApiAbpApplicationConfigurationResponse} from "@ayasofyazilim/core-saas/AccountService";
 
-export const { handlers, auth, signIn, signOut } = NextAuth({
+export const {handlers, auth, signIn, signOut} = NextAuth({
   providers: [
     Credentials({
       credentials: {
@@ -32,14 +28,10 @@ export const { handlers, auth, signIn, signOut } = NextAuth({
           if (signInResponse.error_description) {
             return authorizeError(signInResponse.error_description);
           }
-          const { access_token, refresh_token, expires_in } = signInResponse;
+          const {access_token, refresh_token, expires_in} = signInResponse;
           const expiration_date = expires_in * 1000 + Date.now();
 
-          const user_data = await getUserData(
-            access_token,
-            refresh_token,
-            expiration_date,
-          );
+          const user_data = await getUserData(access_token, refresh_token, expiration_date);
           return user_data;
         } catch (error) {
           return authorizeError(JSON.stringify(error));
@@ -55,18 +47,19 @@ export const { handlers, auth, signIn, signOut } = NextAuth({
     strategy: "jwt",
   },
   callbacks: {
-    signIn({ user }) {
+    signIn({user}) {
       if (user.userName) {
         return true;
       }
       return false;
     },
-    async session({ session, token }) {
+    async session({session, token}) {
       if (token?.user) {
         const user = token?.user as AdapterUser & MyUser;
         if (user.expiration_date < Date.now()) {
-          const { access_token, refresh_token, expires_in } =
-            await fetchNewAccessTokenByRefreshToken(user.refresh_token || "");
+          const {access_token, refresh_token, expires_in} = await fetchNewAccessTokenByRefreshToken(
+            user.refresh_token || "",
+          );
 
           if (access_token && refresh_token) {
             user.access_token = access_token;
@@ -78,11 +71,11 @@ export const { handlers, auth, signIn, signOut } = NextAuth({
       }
       return session;
     },
-    authorized: async ({ auth }) => {
+    authorized: async ({auth}) => {
       // Logged in users are authenticated, otherwise redirect to login page
       return !!auth;
     },
-    async jwt({ token, user }) {
+    async jwt({token, user}) {
       if (user) {
         token.user = user;
       }
