+++ packages/utils/policies/action-policy.tsx
@@ -1,14 +1,9 @@
 "use client";
 
-import { Policy } from "./types";
+import {Policy} from "./types";
 
-export default function isActionGranted(
-  requiredPolicies: Policy[],
-  grantedPolicies: Record<Policy, boolean> | undefined,
-) {
-  const missingPolicies = requiredPolicies.filter(
-    (policy) => !grantedPolicies?.[policy],
-  );
+export function isActionGranted(requiredPolicies: Policy[], grantedPolicies: Record<Policy, boolean> | undefined) {
+  const missingPolicies = requiredPolicies.filter((policy) => !grantedPolicies?.[policy]);
   if (missingPolicies.length > 0) {
     return false;
   }
